requirements.txt
drawsvg~=2.4.0
opencv-python~=4.10.0.84
numpy~=1.26.4
tqdm~=4.67.1
diffusers~=0.32.2
torch~=2.7.0
transformers~=4.49.0
accelerate~=1.4.0
CairoSVG~=2.8.2
matplotlib~=3.10.0
pillow~=10.4.0
open3d~=0.19.0